requirements.txt
numpy==1.18.1
opencv-python==4.2.0.32
tensorflow-gpu==2.0.1
segmentation-models==1.0.1
